// src/app/modules/core/direct-business/models/doiModel.ts
export class PolicyProposalOffer {
    proposerName: string;
    district: string;
    taluka: string;
    referenceNo: string;
    proposalDate: Date;
}
// ...
export class TalukaList {
    value: string;
    district: string;
    viewValue: string;
}